feat: rename entrypoint to cli
