+++ tests/e2e/test_common_command.py
@@ -1,14 +1,14 @@
 from click.testing import CliRunner
 
-from terranova import terranova
-from terranova.terranova import main
+from terranova import cli
+from terranova.cli import main
 from tests.e2e.conftest import assert_result
 
 
 def test_version(runner: CliRunner) -> None:
     result = runner.invoke(main, args=["--version"])
     stdout, _ = assert_result(result)
-    assert f"terranova, version {terranova.__version__}" in stdout
+    assert f"terranova, version {cli.__version__}" in stdout
 
 
 def test_with_no_command(runner: CliRunner) -> None:
